requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyMySQL==1.0.2
Werkzeug==2.2.3
zipp==3.15.0
python-dotenv==1.0.0